requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blobfile==2.1.1
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
einops==0.8.0
fairscale==0.4.13
filelock==3.15.4
fire==0.6.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
huggingface-hub==0.23.5
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.5
-e git+https://github.com/minjunes/llama3.git@e14519ce6a4296858940c68a37bded0a929481d1#egg=llama3
lxml==4.9.4
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycryptodomex==3.20.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
six==1.16.0
sympy==1.13.0
termcolor==2.4.0
tiktoken==0.4.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
-e git+https://github.com/minjunes/vit-pytorch.git@e3256d77cdc890cce68f717b3633e7e4180bd829#egg=vit_pytorch
xxhash==3.4.1
yarl==1.9.4
